backend/requirements.txt
# StuDenTools API Dependencies

# Core Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# PDF Processing
PyPDF2>=3.0.0
# IMPORTANT: opencv-python-headless MUST be installed before pdf2docx
# to prevent pdf2docx from pulling in opencv-python (which needs X11 libs)
opencv-python-headless>=4.8.0
pdf2docx>=0.5.6
pdfplumber>=0.10.0

# Image Processing
Pillow>=10.0.0

# AI/ML - Paraphraser (via OpenRouter)
openai>=1.0.0

# Data Validation
pydantic>=2.0.0

# Email
python-dotenv>=1.0.0
httpx>=0.25.0

# Excel Processing
pandas>=2.0.0
openpyxl>=3.1.0

# Rate Limiting
slowapi>=0.1.9
